### 2024-02-29 #6429 @editorialbot
[PRE REVIEW]: TransitionInTimeSeries.jl: A performant, extensible and reliable software for reproducible detection and prediction of transitions in timeseries
**Submitting author:** @JanJereczek (<a href="http://orcid.org/0000-0003-2213-0423">Jan Swierczek-Jereczek</a>)
**Repository:** https://github.com/JuliaDynamics/TransitionsInTimeseries.jl
**Branch with paper.md** (empty if default branch):
**Version:** v1.0.0
