+++ paper/paper.bib
@@ -5,7 +5,6 @@
 	issn = {02773791},
 	url = {https://linkinghub.elsevier.com/retrieve/pii/S027737910900208X},
 	doi = {10.1016/j.quascirev.2009.06.013},
-	abstract = {The EPICA ice core from Dome C extends 3259 m in depth, and encompasses 800 ka of datable and sequential ice. Numerous chemical species have been measured along the length of the cores. Here we concentrate on interpreting the main low-resolution patterns of major ions. We extend the published record for non-sea-salt calcium, sea-salt sodium and non-sea-salt sulfate ﬂux to 800 ka. The non-sea-salt calcium record conﬁrms that terrestrial dust originating from South America closely mirrored Antarctic climate, both at orbital and millennial timescales. A major cause of the main trends is most likely climate in southern South America, which could be sensitive to subtle changes in atmospheric circulation. Seasalt sodium also follows temperature, but with a threshold at low temperature. We re-examine the use of sodium as a sea ice proxy, concluding that it is probably reﬂecting extent, with high salt concentrations reﬂecting larger ice extents. With this interpretation, the sodium ﬂux record indicates low ice extent operating as an ampliﬁer in warm interglacials. Non-sea-salt sulfate ﬂux is almost constant along the core, conﬁrming the lack of change in marine productivity (for sulfur-producing organisms) in the areas of the Southern Ocean contributing to the ﬂux at Dome C. For the ﬁrst time we also present long records of reversible species such as nitrate and chloride, and show that the pattern of post-depositional losses described for shallower ice is maintained in older ice. It appears possible to use these concentrations to constrain snow accumulation rates in interglacial ice at this site, and the results suggest a possible correction to accumulation rates in one early interglacial. Taken together the chemistry records offer a number of constraints on the way the Earth system combined to give the major climate ﬂuctuations of the late Quaternary period.},
 	language = {en},
 	number = {1-2},
 	urldate = {2022-06-14},
@@ -23,7 +22,6 @@
 	issn = {1932-6203},
 	url = {https://dx.plos.org/10.1371/journal.pone.0041010},
 	doi = {10.1371/journal.pone.0041010},
-	abstract = {Many dynamical systems, including lakes, organisms, ocean circulation patterns, or financial markets, are now thought to have tipping points where critical transitions to a contrasting state can happen. Because critical transitions can occur unexpectedly and are difficult to manage, there is a need for methods that can be used to identify when a critical transition is approaching. Recent theory shows that we can identify the proximity of a system to a critical transition using a variety of so-called ‘early warning signals’, and successful empirical examples suggest a potential for practical applicability. However, while the range of proposed methods for predicting critical transitions is rapidly expanding, opinions on their practical use differ widely, and there is no comparative study that tests the limitations of the different methods to identify approaching critical transitions using time-series data. Here, we summarize a range of currently available early warning methods and apply them to two simulated time series that are typical of systems undergoing a critical transition. In addition to a methodological guide, our work offers a practical toolbox that may be used in a wide range of fields to help detect early warning signals of critical transitions in time series data.},
 	language = {en},
 	number = {7},
 	urldate = {2022-09-29},
@@ -42,10 +40,6 @@
 	issn = {0027-8424, 1091-6490},
 	url = {https://pnas.org/doi/full/10.1073/pnas.2106140118},
 	doi = {10.1073/pnas.2106140118},
-	abstract = {Significance
[…]
