### 2024-03-06 80db37a
Minors changes for JOSS paper (#76)
* ignore paper

* remove csic from affiliations

* remove abstracts

* fix link to docs

* replace Figure kwarg `resolution` by  `size`
