+++ codecov.yml
@@ -4,3 +4,4 @@ ignore:
   - "src/misc/load_data.jl"
   - "src/misc/params.jl"
   - "src/visualizations.jl"
+  - "paper/*"
