### 2024-03-11 #6429 @lrnv
:wave: @sgeorge91 and @felixcremer, would any of you be willing to review this submission for JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html
