### 2024-03-11 #6429 @lrnv
@felixcremer The expected time line is about 6 to 8 weeks but can be extended on demand. We aim at publication after 3 to 4 months. Anyway, if you need more time in the middle, you can always ask nicely and usually the editor (that is, me) will give you more time ;)
