### 2024-03-13 #6464 @Datseris
Hello @lrnv , thank you for coordinating this review.

We noticed that the _title_ of our paper has a typo. The software is called TransitionsInTimeseries.jl but the title of the paper states TransitionInTimeseries.jl (notice the singular Transition). We hope it is possible to alter the paper title after the review process has started?
