### 2024-03-13 #6464 @lrnv
I think you can simply change the title in the paper.md file and that will be enough, since (at the publishing time), this is the fil that will be used everywhere. Let me come back to you with confirmation.
