### 2024-03-13 #6464 @lrnv
So confirmed, it is enough. Regenerate the paper after that and I'll edit this issue title accordingly.
