### 2024-03-13 45e4185
include tests for slope change
