+++ previews/PR49/tutorial.jl
@@ -1,0 +1,263 @@
+#=
+
+# Tutorial
+
+## [Workflow] (@id workflow)
+
+Computing transition indicators consists of the following steps:
+
+1. Doing any preprocessing of raw data first, such as detrending (_not part of TransitionsInTimeseries.jl_). This yields the **input timeseries**.
+2. Estimating the timeseries of an indicator by sliding a window over the input timeseries.
+3. Computing the changes of the indicator by sliding a window over its timeseries.
+4. Generating many surrogates that preserve important statistical properties of the original timeseries.
+5. Performing step 2 and 3 for the surrogate timeseries.
+6. Checking whether the indicator change timeseries of the real timeseries shows a significant feature (trend, jump or anything else) when compared to the surrogate data.
+
+These steps are illustrated one by one in the tutorial below, and then summarized in the convenient API that TransitionsInTimeseries.jl exports.
+
+## [Tutorial -- Educational] (@id example_stepbystep)
+
+### Raw input data
+
+Let us load data from a bistable nonlinear model subject to noise and to a gradual change of the forcing that leads to a transition. Furthermore, we also load data from a linear model, which is by definition monostable and therefore incapable of transitioning. This is done to control the rate of false positives, a common problem that can emerge when looking for transition indicators. The models are governed by:
+
+```math
+\dfrac{\mathrm{d}x_{l}}{\mathrm{d}t} = - x_{l} - 1 + f(t) + n(t) \\
+\dfrac{\mathrm{d}x_{nl}}{\mathrm{d}t} = - x_{nl}^3 + x_{nl} + f(t) + n(t)
+```
+
+with $x_{l}$ the state of the linear model, $x_{nl}$ the state of the bistable model, $f$ the forcing and $n$ the noise. For $f=0$ they both display an equilibrium point at $x=-1$. However, the bistable model also displays a further equilibrium point at $x=1$. Loading (and visualizing with [Makie](https://docs.makie.org/stable/)) such prototypical data to test some indicators can be done by simply running:
+=#
+
+using TransitionsInTimeseries, CairoMakie
+
+t, x_linear, x_nlinear = load_linear_vs_doublewell()
+fig, ax = lines(t, x_linear)
+lines!(ax, t, x_nlinear)
+ax.title = "raw data"
+fig
+
+#=
+### Preprocessing
+
+!!! note "Not part of TransitionsInTimeseries.jl"
+    Any timeseries preprocessing, such as the de-trending step we do here,
+    is not part of TransitionsInTimeseries.jl and is the responsibility of the researcher.
+
+
+The nonlinear system clearly displays a transition between two stability regimes. To forecast such transition, we analyze the fluctuations of the timeseries around the attractor, assumed to be tracked. Therefore, a detrending step is needed - here simply obtained by building the difference of the timeseries with lag 1.
+=#
+
+x_l_fluct = diff(x_linear)
+x_nl_fluct = diff(x_nlinear)
+tfluct = t[2:end]
+
+fig, ax = lines(tfluct, x_l_fluct)
+lines!(ax, tfluct, x_nl_fluct .+ 0.05)
+ax.title = "input timeseries"
+fig
+
+#=
+At this point, `x_l_fluct` and `x_nl_fluct` are considered the **input timeseries**.
+
+!!! info "Detrending in Julia"
+    Detrending can be performed in many ways. A wide range of Julia packages exists to perform smoothing such as [Loess.jl](https://github.com/JuliaStats/Loess.jl) or [DSP.jl](https://docs.juliadsp.org/latest/contents/). There the detrending step consists of subtracting the smoothed signal from the original one.
+
+### Indicator timeseries
+
[…]

[…]
