+++ v0.1.1/examples/do-events.jl
@@ -1,0 +1,318 @@
+#=
+# Dansgaard-Oescher events and Critical Slowing Down
+
+The $\delta^{18}O$ timeseries of the North Greenland Ice Core Project
+([NGRIP](https://en.wikipedia.org/wiki/North_Greenland_Ice_Core_Project)) are,
+to this date, the best proxy record for the Dansgaard-Oeschger events
+([DO-events](https://en.wikipedia.org/wiki/Dansgaard%E2%80%93Oeschger_event)).
+DO-events are sudden warming episodes of the North Atlantic, reaching 10 degrees
+of regional warming within 100 years. They happened quasi-periodically over the
+last glacial cycle due to transitions between strong and weak states of the Atlantic
+Meridional Overturning Circulation and might be therefore be the most prominent
+examples of abrupt transitions in the field of climate science. We here propose
+to hindcast these events by applying the theory of Critical Slowing Down (CSD)
+on the NGRIP data, which can be found [here](https://www.iceandclimate.nbi.ku.dk/data/)
+in its raw format. This analysis has already been done in [boers-early-warning-2018](@cite)
+and we here try to reproduce Figure 2.d-f.
+
+## Preprocessing NGRIP
+
+Data pre-processing is not part of TransitionsInTimeseries.jl,
+but a step the user has to do before using the package.
+To present an example with a complete scientific workflow,
+we will showcase typical data pre-processing here, that consist of the following steps:
+1. Load the data, reverse and offset it to have time vector = time before 2000 AD.
+2. Filter non-unique points in time and sort the data.
+3. Regrid the data from uneven to even sampling.
+
+The time and $\delta^{18}O$ vectors resulting from the $i$-th preprocessing step are
+respectively called $t_i$ and $x_i$. The final step consists in obtaining a residual
+$r$, i.e. the fluctuations of the system around the attractor, which, within the CSD
+theory, is assumed to be tracked. Over this example, it will appear that the
+convenience of TransitionsInTimeseries.jl leads the bulk of the code to be written
+for plotting and preprocessing.
+
+### Step 1:
+=#
+
+using DelimitedFiles, Downloads
+
+function load_ngrip()
+    tmp = Base.download("https://raw.githubusercontent.com/JuliaDynamics/JuliaDynamics/"*
+        "master/timeseries/NGRIP.csv")
+    data, labels = readdlm(tmp, header = true)
+    return reverse(data[:, 1]) .- 2000, reverse(data[:, 2]) # (time, delta-18-0) vectors
+end
+
+t1, x1 = load_ngrip()
+
+#=
+### Step 2
+=#
+
+uniqueidx(v) = unique(i -> v[i], eachindex(v))
+function keep_unique(t, x)
+    unique_idx = uniqueidx(t)
+    return t[unique_idx], x[unique_idx]
+end
+
+function sort_timeseries!(t, x)
+    p = sortperm(t)
+    permute!(t, p)
+    permute!(x, p)
+    return nothing
+end
+
+t2, x2 = keep_unique(t1, x1)
+sort_timeseries!(t2, x2)
+
+#=
+### Step 3
+=#
+
+using BSplineKit
+
+function regrid2evensampling(t, x, dt)
+    itp = BSplineKit.interpolate(t, x, BSplineOrder(4))
+    tspan = (ceil(minimum(t)), floor(maximum(t)))
+    t_even = collect(tspan[1]:dt:tspan[2])
+    x_even = itp.(t_even)
+    return t_even, x_even
+end
+
+dt = 5.0        # dt = 5 yr as in (Boers 2018)
+t3, x3 = regrid2evensampling(t2, x2, dt)
+
+#=
+### Step 4
+
+For the final step we drop the indices:
+=#
+
+using DSP
+
+function chebyshev_filter(t, x, fcutoff)
[…]
