### 2024-04-26 #6464 @sgeorge91
> @Datseris , I am not sure if its something local going on my system, but I seem to have some trouble running some of the functions defined in analysis. In particular, I run into an error when I try running estimate_changes(...). ERROR: UndefVarError: `estimate_changes` not defined Julia isn't really my strong point, so apologies if this is something trivial. But I've tried uninstalling and reinstalling the package, and it leads to the same error.
>
> With respect to the paper, I was largely happy with everything in it. I had a two small suggestions for improvements of the paper/package.
>
>     1. The statement of need: While the current statement covers most of what is needed, I thought an important aspect to mention is the lack of support for user defined functions in existing packages. I thought that this was an important improvement in the present package (as mentioned in the section on 'comparison to already existing alternatives'), and could be mentioned earlier in the manuscript, possibly in the statement of need.
>
>     2. I noticed that the p-value of the Kendall tau itself is not used for determining significance. This makes sense, since the p-value is often severely underestimated in autocorrelated data. I just wanted to point out an extension of the Mann Kendall trend test which corrects for the assumption of independence while using sliding windows (Hamed, K. H., & Rao, A. R. (1998). A modified Mann-Kendall trend test for autocorrelated data. Journal of hydrology, 204(1-4), 182-196). May be something to consider implementing in a future version

Never mind. I realized I was installing a wrong version. It works fine now, and I am happy with the paper and the package. Minor ideas for changes are mentioned above.
Thanks!
