### 2024-05-07 JuliaDynamics/TransitionsInTimeseries.jl#79 @felixcremer
Code example in JOSS paper does not work
Trying the code example in the JOSS paper I first got an `residual` not defined error, but when I replace residual with `data` I get the following error:

```julia
julia> flags = significant_transitions(results, signif)
ERROR: MethodError: no method matching length(::Symbol)

Closest candidates are:
  length(::Combinatorics.Partition)
   @ Combinatorics ~/.julia/packages/Combinatorics/Udg6X/src/youngdiagrams.jl:8
  length(::DataStructures.EnumerateAll)
   @ DataStructures ~/.julia/packages/DataStructures/95DJa/src/multi_dict.jl:96
  length(::Pkg.Types.Manifest)
   @ Pkg ~/.julia/juliaup/julia-1.10.2+0.x64.linux.gnu/share/julia/stdlib/v1.10/Pkg/src/Types.jl:315
  ...

Stacktrace:
 [1] sanitycheck_tail(tail::Symbol, n_ind::Int64)
   @ TransitionsInTimeseries ~/Documents/TransformationsInTimeseriesReview/dev/TransitionsInTimeseries/src/significance/surrogates_significance.jl:177
 [2] significant_transitions(res::SegmentedWindowResults{…}, signif::SurrogatesSignificance{…})
   @ TransitionsInTimeseries ~/Documents/TransformationsInTimeseriesReview/dev/TransitionsInTimeseries/src/significance/surrogates_significance.jl:120
 [3] top-level scope
   @ REPL[141]:1
Some type information was truncated. Use `show(err)` to see complete types.
```

This is part of the JOSS review https://github.com/openjournals/joss-reviews/issues/6464
