### 2024-05-07 JuliaDynamics/TransitionsInTimeseries.jl#81 @felixcremer
Cannot reproduce benchmarks ewstools_tuto-1.csv missing
I can not reproduce the benchmarks of TransitionsInTimeseries.jl because the file `ewstools_tuto-1.csv` is missing.

Could you please provide this file in the paper/code folder?
This is part of the JOSS review https://github.com/openjournals/joss-reviews/issues/6464
