+++ ext/TransitionVisualizations.jl
@@ -51,8 +51,6 @@ function TransitionsInTimeseries.plot_indicator_changes(res::SegmentedWindowResu
     fig, n = init_rowwise_visualisation(res, colors, indicator_names,
         chametric_names, accent_linewidth)
     for k in eachindex(res.t_indicator)
-        Makie.lines!(contents(fig[1, 1])[1], res.t[k], res.x[k], color = colors[1],
-            linewidth = accent_linewidth)
         lineplot_metrics!(fig, n, res.config, res.t_indicator[k], res.x_indicator[k],
             res.t_change[k], res.x_change[k, :], colors, accent_linewidth)
     end
@@ -65,17 +63,17 @@ function init_rowwise_visualisation(res, colors, indicator_names,
     chametric_names, accent_linewidth)
 
     fig = Makie.Figure(size = (700, 450), fontsize = 12)
-    rlabels = vcat([""], indicator_names)
-    llabels = vcat(["Input"], chametric_names)
+    llabels = vcat(["Input"], indicator_names)
+    rlabels = vcat([""], chametric_names)
     n = length(llabels)
 
     # rowaspect = 5
-    raxs = [Makie.Axis(fig[i, 1], ylabel = rlabels[i],
-        xticklabelsvisible = false, yaxisposition = :right, ygridvisible = false,
-        ylabelcolor = colors[2], yticklabelcolor = colors[2]) for i in 1:n]
     laxs = [Makie.Axis(fig[i, 1], ylabel = llabels[i],
         xticklabelsvisible = false, ygridvisible = false,
         ylabelcolor = colors[1], yticklabelcolor = colors[1]) for i in 1:n]
+    raxs = [Makie.Axis(fig[i, 1], ylabel = rlabels[i],
+        xticklabelsvisible = false, yaxisposition = :right, ygridvisible = false,
+        ylabelcolor = colors[2], yticklabelcolor = colors[2]) for i in 1:n]
     Makie.linkxaxes!(laxs..., raxs...)
 
     hidedecorations!(raxs[1])
@@ -90,7 +88,7 @@ function init_rowwise_visualisation(res, colors, indicator_names,
     labels = ["original signal", "surro signals"]
     width = 0.5
     if length(res.t_indicator[1]) > 1
-        elements = vcat(elements, [MarkerElement(marker = :circle, color = colors[1],
+        elements = vcat(elements, [MarkerElement(marker = :circle, color = colors[2],
             strokecolor = :transparent, markersize = ms) for ms in [10, 5]])
         labels = vcat(labels, ["original change metric", "surro change metrics"])
         width = 1
@@ -106,14 +104,14 @@ function lineplot_metrics!(fig, n, config, t_ind, x_ind, t_cha, x_cha,
     for i in 2:n
         j = i-1
         if !isnothing(config.indicators)
-            Makie.lines!(contents(fig[i, 1])[2], t_ind, x_ind[:, j], color = colors[2],
+            Makie.lines!(contents(fig[i, 1])[1], t_ind, x_ind[:, j], color = colors[1],
                 linewidth = accent_linewidth)
         end
         if length(t_cha) > 1
-            lines!(contents(fig[i, 1])[1], t_cha, x_cha[:, j], color = colors[1],
+            lines!(contents(fig[i, 1])[2], t_cha, x_cha[:, j], color = colors[2],
                 linewidth = accent_linewidth)
         else
-            Makie.scatter!(contents(fig[i, 1])[1], t_cha, x_cha[j], color = colors[1],
+            Makie.scatter!(contents(fig[i, 1])[2], t_cha, x_cha[j], color = colors[2],
                 markersize = 10)
         end
     end
[…]
