### 2024-05-21 3e66e40
change timing to min (consistent with @belapsed)
