+++ paper/code/figure2.jl
@@ -1,18 +1,22 @@
-using TransitionsInTimeseries, DelimitedFiles, CairoMakie, Random
+using CairoMakie
+using DelimitedFiles
+using Random
+using TransitionsInTimeseries
 
 coefficient_of_variation(x) = std(x) / mean(x)
 
 function main()
-    x = readdlm("ewstools-tuto-1.csv", ',')[:, end]
+    # Run ewstools-tuto-1.csv first to generate ewstools_ricker.csv and ewstools_perfo.csv
+    x = readdlm("../data/ewstools_ricker.csv", ',')[:, end]
+    x = x[isnan.(x) .== 0]
     t = eachindex(x)
 
     # Choose the indicators and how to measure their change over time
     indicators = (var, coefficient_of_variation, skewness, kurtosis,
         ar1_whitenoise, LowfreqPowerSpectrum())
     stride = [1, 1, 1, 1, 1, 40]
-    n, m = 100, length(indicators)
+    m = length(indicators)
     t_elapsed = zeros(m+2)
-
 
     for (i, ind) in enumerate(indicators)
         # Build configuration with adequate parameters of the sliding window
@@ -20,50 +24,37 @@ function main()
             width_ind = length(x) ÷ 2, stride_ind = stride[i], whichtime = last,
             min_width_cha = 1)
 
-        t0 = time()
-        for i in 1:n
-            # Compute the metrics over sliding windows and their significance
-            results = estimate_changes(config, x, t)
-        end
-        t_elapsed[i] = (time() - t0) / 2
+        t_elapsed[i] = @belapsed estimate_changes($config, $x, $t)
     end
 
     config = SegmentedWindowConfig((nothing, nothing), (kendalltau, kendalltau), [t[1]], [t[end]];
         width_ind = length(x) ÷ 2, stride_ind = 1, whichtime = last, min_width_cha = 1)
-    t0 = time()
-    for i in 1:n
-        results = estimate_changes(config, x, t)
-    end
-    t_elapsed[m+1] = (time() - t0) / 2
+    t_elapsed[m+1] = @belapsed estimate_changes($config, $x, $t)
 
     sgen = surrogenerator(x, BlockShuffle(), Xoshiro(1995))
-    t0 = time()
-    for i in 1:n
-        s = sgen()
-    end
-    t_elapsed[m+2] = time() - t0
+    t_elapsed[m+2] = @belapsed $sgen()
 
     return t_elapsed
 end
 
-t_tt = main()
-t_et = [0.03840542, 0.05554581, 0.03895116, 0.04029274, 7.96556187,
-    2.73067856, 0.39529872, 0.02751493]
-
-# [0.04681492, 8.13679838, 0.04035759, 0.09219241]
-inds = eachindex(t_et)
+t_transitionsintimeries = main()
+t_ewstools = readdlm("../data/ewstools_perfo.csv", ',')[:, end]
+inds = eachindex(t_ewstools)
 w = 0.4
 
-fig, ax = barplot(inds .- 0.5*w, t_et, label = L"ewstools $\,$", width = w,
-    fillto = 1e-5)
-barplot!(ax, inds .+ 0.5*w, t_tt, label = L"TransitionsInTimeseries.jl $\,$",
-    width = w, fillto = 1e-5)
+fig = Figure(resolution = (600, 400))
+ax = Axis(fig[1, 1])
+ylims!(ax, 1e-7, 0.1)
+barplot!(ax, inds .- 0.5*w, t_ewstools, label = L"ewstools $\,$", width = w,
+    fillto = 1e-8)
[…]
