+++ paper/code/Project.toml
@@ -1,4 +1,4 @@
 [deps]
+BenchmarkTools = "6e4b80f9-dd63-53aa-95a3-0cdb28fa8baf"
 CairoMakie = "13f3f980-e62b-5c42-98c6-ff1f3baf88f0"
 DelimitedFiles = "8bb1440f-4735-579b-a4ab-409b98df4dab"
-TransitionsInTimeseries = "5f5b98ec-1183-43e0-887a-12fdc55c52f7"
