### 2024-05-21 e7c7e63
add extensibility earlier + modify description of performance
