### 2024-05-22 d040fae
generalise to OffsetArrays
