### 2024-05-22 cb0d387
Addressing the reviews for JOSS submission (#83)
* permute left and right axis to ease reading

* write output to csv

* correct examples

* csv outputs to data folder

* read output from new csv file

* change timing to min (consistent with @belapsed)

* use @belapsed

* change csv name in comment

* update figures

* add extensibility earlier + modify description of performance

* specify ewstools version

* generalise to OffsetArrays

* remove unused packages and add OffsetArrays
