+++ paper/paper.md
@@ -44,12 +44,15 @@ work predicting a collapse of the Atlantic Meridional Overturning Circulation be
 and 2095 has led to no less than 870 news outlets and 4100 tweets [@ditlevsen_warning_2023],
 largely because of the substantial implications of such a collapse for human societies.
 A common concern in the scientific community is that published work on the topic is difficult
-to reproduce, despite the impact it implies for humanity.
-This can be largely addressed by a unifying software that is accessible, performant,
-reproducible, reliable and extensible. Such a software does not exist yet, but here
-we propose TransitionsInTimeseries.jl to fill this gap.
-We believe this is a major step towards establishing a software as standard, widely used
-by academics working on transitions in timeseries.
+to reproduce, despite the impact it implies for humanity. Additionally, existing softwares
+[@dakos_methods_2012, @bury_ewstools_2023] are well suited to apply some state-of-the-art
+techniques but lack the possibility of being easily extended by the users, thus making them
+unsuited for some research task.
+Both of these issues can be largely addressed by a unifying software that is accessible,
+performant, reproducible, reliable and extensible. Such a software does not exist yet, but
+here we propose TransitionsInTimeseries.jl to fill this gap. We believe this is a major step
+towards establishing a software as standard, widely used by academics working on transitions
+in timeseries.
 
 # TransitionsInTimeseries.jl
 
@@ -77,8 +80,11 @@ surrogates, corresponding to a p-value $p<0.05$. All these steps can be performe
 visualisation of the results within a few lines only:
 
 ```julia
+using TransitionsInTimeseries, CairoMakie, Random
+
 # Loading and preprocessing the data needs to be done by the user
 time, data = load_data()
+residual = data - filt(data)
 
 # Choose the indicators and how to measure their change over time
 indicators = (var, ar1_whitenoise)
@@ -89,8 +95,8 @@ config = SegmentedWindowConfig(indicators, change_metrics, [time[1]], [time[end]
     width_ind = length(residual) ÷ 2, whichtime = last, min_width_cha = 100)
 
 # Compute the metrics over sliding windows and their significance
-results = estimate_changes(config, data, time)
-signif = SurrogatesSignificance(n = 1000, tail = :right, rng = Xoshiro(1995))
+results = estimate_changes(config, residual, time)
+signif = SurrogatesSignificance(n = 1000, tail = (:right, :right), rng = Xoshiro(1995))
 flags = significant_transitions(results, signif)
 
 # Visualize the results
@@ -118,7 +124,7 @@ TransitionsInTimeseries.jl is written in Julia, which offers both a simple synta
 performance. Additionally, all performance-relevant steps have been optimized and
 parallelized when possible, as, for instance, the significance testing relying on surrogates.
 In the final section of this article, we present a comparison to `ewstools`, showing that
-TransitionsInTimeseries.jl offers a significant speed-up in all the studied cases.
+TransitionsInTimeseries.jl offers a significant speed-up in most the studied cases.
 
 ## Reproducibility
 
@@ -173,17 +179,15 @@ thus offering optimized routines with numerous surrogate types.
 ### Choosing a pipeline
 
 TransitionsInTimeseries.jl covers methods for prediction as well as detection of transitions,
-which is unprecedented to our knowledge. This relies on the definition of different analysis pipelines, which
-consist in a `ChangesConfig` determining the behavior of `estimate_changes` via
-multiple dispatch. For instance, a detection
-task can be performed by replacing the `SegmentedWindowConfig` by a `SlidingWindowConfig`
-in the code above:
[…]
