+++ test/timeseries.jl
@@ -1,12 +1,14 @@
-using TransitionsInTimeseries, Test
+using TransitionsInTimeseries, Test, OffsetArrays
 
 @testset "spacing" begin
     dt = rand()
     x_equispaced = range( 0.0, step = dt, stop = 100)
+    x_equispaced_offset = OffsetVector([1:10...], -1:8)
     x_nonequispaced = cumsum( rand(1000) )
     @test isequispaced(x_equispaced)
     @test !isequispaced(x_nonequispaced)
     @test equispaced_step(x_equispaced) ≈ dt
+    @test equispaced_step(x_equispaced_offset) ≈ 1
 end
 
 #=
