+++ previews/PR39/search_index.js
@@ -1,0 +1,3 @@
+var documenterSearchIndex = {"docs":
+[{"location":"api/#API","page":"API","title":"API","text":"","category":"section"},{"location":"api/#Main-analysis-functions","page":"API","title":"Main analysis functions","text":"","category":"section"},{"location":"api/","page":"API","title":"API","text":"TransitionsSurrogatesConfig\nestimate_transitions\nTransitionsResults\ntransition_flags","category":"page"},{"location":"api/#TransitionsInTimeseries.TransitionsSurrogatesConfig","page":"API","title":"TransitionsInTimeseries.TransitionsSurrogatesConfig","text":"TransitionsSurrogatesConfig(indicators, change_metrics [, surrogate]; kwargs...)\n\nA configuration struct for TransitionsInTimeseries.jl that can be given to estimate_transitions. It contains all information necessary to perform the basic Workflow of the package to detect significant transitions in the input timeseries, using the method of surrogate testing to quantify significance:\n\nEstimate the timeseries of an indicator by sliding a window over the input timeseries.\nEstimate changes for an indicator by sliding a window over its timeseries.\nGenerate many surrogates that preserve important statistical properties of the original input timeseries.\nPerform steps 1 and 2 for the surrogate timeseries (and for all provided indicators).\nEstimate when an indicators timeseries shows significant change (trend, jump or anything else) when compared to the surrogate timeseries when compared to the surrogate data. Significance is estimated from the p-values of the real data vs surrogate data.\n\nArguments\n\nindicators::AbstractVector{<:Function} a vector of indicators. Some indicators typically used in the literature are listed in the documentation section on indicators. The analysis is performed efficiently for all indicators given.\nchange_metrics change metrics corresponding to the given indicators. If given a function, the same function is used for all indicators. Otherwise, it should be a vetor of functions of the same size as indicators, using change metric for its corresponding indicator. Some change metrics typically used in the literature are listed in the documentation section on change metrics.\nsurrogate::Surrogate the method to use to generate surrogates of the input timeseries. This is an optional argument that defaults to RandomFourier(), see Surrogates for alternative options.\n\nBoth indicators and change metrics are generic Julia functions that input an x::AbstractVector and output an s::Real. Any appropriate function may be given and see making custom indicators/change metrics in the documentation for more information.\n\nKeyword arguments\n\nwidth_ind::Int=100, stride_ind::Int=1: width and stride given to the WindowViewer to compute the indicator from the input timeseries.\nwidth_cha::Int=50, stride_cha::Int=1: width and stride given to the WindowViewer to compute the change metric timeseries from the indicator timeseries.\nwhichtime = midpoint: The time vector corresponding to the indicators / change metric timeseriesm is obtained from t using the keyword whichtime. Options include:\nlast: use the last timepoint of each window\nmidpoint: use the mid timepoint of each time window\nfirst: use first timepoint of each window\nIn fact, the indicators time vector is computed simply via\nt_indicator = windowmap(whichtime, t; width_ind, stride_ind)\nt_change = windowmap(whichtime, t_indicator; width_cha, stride_cha)\nso any other function of the time window may be given to extract the time point itself, such as mean or median.\nn_surrogates::Int = 10_000: how many surrogates to create.\nrng::AbstractRNG = Random.de
[…]
