+++ previews/PR45/examples.jl
@@ -1,0 +1,192 @@
+# # Examples for TransitionsInTimeseries.jl
+
+# ## Permutation entropy for dynamic regime changes
+
+# Permutation entropy is used frequently to detect a transition between
+# one dynamic regime to another. It is useful when the mean and std. of the
+# timeseries values are very similar between the two regimes, which would mean that
+# common distribution-based indicators, or common critical-slowing-down based indicators,
+# would fail.
+
+# This example will also explore different ways to test for significance that
+# are arguably better suitable in such an application than the [Tutorial](@ref)'s default
+# of significance via random Fourier surrogates.
+
+# ### Logistic map timeseries
+
+# A simple example of this is transitions from periodic to weakly chaotic to chaotic
+# motion in the logistic map. First, let's generate a timeseries of the logistic map
+
+using DynamicalSystemsBase
+using CairoMakie
+
+logistic_rule(u, p, t) = @inbounds SVector(p[1]*u[1]*(1 - u[1]))
+ds = DeterministicIteratedMap(logistic_rule, [0.5], [1.0])
+
+r1 = 3.83
+r2 = 3.86
+N = 2000
+rs = range(r1, r2; length = N)
+x = zeros(N)
+for (i, r) in enumerate(rs)
+    set_parameter!(ds, 1, r)
+    step!(ds)
+    x[i] = current_state(ds)[1]
+end
+
+# Plot it, using as time the parameter value (they coincide)
+fig, ax = lines(rs, x; linewidth = 0.5)
+ax.xlabel = "r (time)"
+ax.ylabel = "x"
+fig
+
+# In this example there is a rather obvious transition to strongly
+# chaotic motion at r ≈ 3.855. However, there is also a subtle transition
+# to weak chaos at r ≈ 3.847. This transition is barely visible in the
+# timeseries, and in fact many of the timeseries statistical properties remain identical.
+
+# ### Using a simpler change metric
+
+# Now, let's compute and various indicators and their changes,
+# focusing on the fourth indicator, the permutation entropy. We use
+# order 4 here, because we know that to detect changes in a period `m` we would need
+# an order ≥ `m+1` permutation entropy.
+
+using TransitionsInTimeseries
+indicators = (var, ar1_whitenoise, permutation_entropy(m = 4))
+indistrings = ("var", "ar1", "pe")
+
+# In this example there is no critical slowing down;
+# instead, there is a sharp transition between periodic and chaotic motion.
+# Hence, we shouldn't be using any trend-based change metrics.
+# Instead, we will use the most basic change metric, [`difference_of_means`](@ref).
+# With this metric it also makes most sense to use as stride half the window width
+metric = difference_of_means
+
+width_ind = N÷100
+width_cha = 20
+stride_cha = 10
+
+config = WindowedIndicatorConfig(indicators, metric;
+    width_ind, width_cha, stride_cha,
+)
+
+results = estimate_indicator_changes(config, x, rs)
+
+# Let's now plot the change metrics of the indicators
+
+function plot_change_metrics()
+    fig, ax = lines(rs, x; axis = (ylabel = "input",), figure = (resolution = (600, 600),))
+    hidexdecorations!(ax; grid = false)
+    ## plot all change metrics
+    for (i, c) in enumerate(eachcol(results.x_change))
+        ax, = scatterlines(fig[i+1, 1], results.t_change, c;
+            axis = (ylabel = indistrings[i],), label = "input"
+        )
+        if i < 3
+            hidexdecorations!(ax; grid = false)
+        else
+            ax.xlabel = "r (time)"
+        end
+    end
+    return fig
+end
+
+fig = plot_change_metrics()
+
+# We already see the interesting results we expect: the permutation entropy shows a
[…]
