### 2024-06-03 #6464 @felixcremer
Is the comparison between python and julia fair in your benchmark?
From my view of the benchmark scripts the python version is using a time series of 1000 steps while the julia benchmarks uses only 861 entries because NaN entries are filtered before computing the timings.
I am not fully sure about this, because it might also be that ewstools is dealing with the NaN entries internally in a similar way as the julia script. I did the benchmark also without the filtering and there are some differences in both directions.
