### 2024-06-03 #6464 @JanJereczek
Hi @felixcremer, that's an important question! `ewstools/core.py` uses following chunk of code for all the rolling window computations:

```
# Get time series data prior to transition
if self.transition:
    df_pre = self.state[self.state.index <= self.transition]
else:
    df_pre = self.state
```

This filters out the points beyond the transition and, for our benchmark example, only 861 entries are used. I think that the benchmark is therefore fair!
