### 2024-07-12 #6464 @JanJereczek
Hi @lrnv, you will find the latest release (v0.1.3) archived at [10.5281/zenodo.12731309](https://zenodo.org/doi/10.5281/zenodo.12731309). I performed all the checks you mentioned but I have a doubt about one of them. @felixcremer made contributions and hence appears as a contributor to the package on both github and zenodo. However, he is obviously not on the author list of the paper. This is not in line with the penultimate point on the checklist you asked us to go through but I think it is correct as it is. Can you help me clarify this?
