### 2024-07-12 #6464 @lrnv
@JanJereczek Before accepting, I do not understand why the figure 1 two lower plots only have data for the right half of the timeframe. Is that intentional ?
