### 2024-07-12 #6464 @JanJereczek
Yes, it is intentional and results from the sliding window approach, which requires points from $t - w$ to $t$ to estimate the statistic at time $t$. The width $w$ of the window is set quite high (c.f. `width_ind = length(residual) \div 2` in the code we show on page 2) and hence you only see timeseries for the statistics in the second half of the time span.
