### 2024-07-12 #6464 @JanJereczek
Indeed! It should be corrected now :) Is there any way I can check if the xml metadata file is generated correctly or do you have to run it yourself @lrnv?
