### 2024-07-13 #6464 @crvernon
:wave: @JanJereczek - great job on this!  We have a few more things to address before moving to accept:

- [ ] Please ensure the authors and the author order match between the paper and the Zenodo archive.  No need to do another release, just edit the Zenodo archive metadata to fix this.

In the paper:
- [ ] LINE 22-23:  "...existing softwares Bury (2023) are well suited to apply..." should be written "...existing software detailed in Bury (2023) are well suited to apply..."
- [ ] LINE 49:  The figure reference is not linking properly due to incorrect formatting.  In your paper.md file, you state the following:  `...shown in [Fig. 1](@figure1) and display,...` which should be `...shown in \autoref{fig:figure1} and display,...`  Then the figure caption can be referenced as:  `![Output of plotting function in usage example.\label{fig:figure1}](figures/figure1.png)`.  This should allow the figure reference in the text to correctly link and format to the figure.  Please change the formatting for figure 2 as well.
- [ ] LINE 127:  "...developed in (Bury et al., 2021), as..." should read "...developed in Bury et al. (2021), as..." You can just remove the brackets around the citation in the text to achieve this.
- [ ] LINE 138:  missing space between "of" and "bifurcations"

No need to create another release.  Just let me know once the changes have been made.  Thanks!
