### 2024-07-13 #6464 @crvernon
Thanks @JanJereczek, the paper looks good. I still need you to address the following:

> Please ensure the authors and the author order match between the paper and the Zenodo archive. No need to do another release, just edit the Zenodo archive metadata to fix this.
