### 2024-07-13 #6464 @JanJereczek
The order matches between both and @felixcremer was added to the author list on zenodo since he made some PRs over the review process. However he is not a co-author of the paper. See the discussion above: https://github.com/openjournals/joss-reviews/issues/6464#issuecomment-2225128873. Sounds good to you?
